TEST/backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
filelock==3.20.0
fsspec==2025.12.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.4
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pgmpy==1.0.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
statsmodels==0.14.6
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.41.0
apscheduler>=3.10,<4.0
